removed some unmaintained code and renamed old InstrumentDescription to mc_config
