+++ examples/display_DL0testbed3_withID.py
@@ -9,7 +9,7 @@ from matplotlib import colors, pyplot as plt
 from matplotlib.backends.backend_pdf import PdfPages
 
 from ctapipe import visualization
-from ctapipe.instrument import InstrumentDescription as ID
+from ctapipe.instrument import mc_config as ID
 from ctapipe.io.hessio import hessio_event_source
 from ctapipe.utils import get_dataset
 from ctapipe.instrument import CameraGeometry
