+++ examples/muon_reco.py
@@ -10,7 +10,7 @@ from ctapipe.reco.cleaning import tailcuts_clean
 from ctapipe import io
 from astropy.coordinates import Angle, AltAz
 from astropy.time import Time
-from ctapipe.instrument import InstrumentDescription as ID
+from ctapipe.instrument import mc_config as ID
 from ctapipe.coordinates import CameraFrame, NominalFrame
 from ctapipe.image.muon.muon_ring_finder import ChaudhuriKunduRingFitter
 from ctapipe.image.muon.muon_integrator import *
